requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cssbeautifier==1.15.4
djlint==1.36.4
dnspython==2.8.0
EditorConfig==0.17.1
email-validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.1
MarkupSafe==3.0.3
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.43
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
